requirements.txt
RPi.GPIO
mqtt-io

 
python3-tzlocal 
python3-sdnotify 
python3-colorama 
python3-unidecode 
python3-apt 
python3-paho-mqtt 
python3-requests
